--- doctor-who-ai/app/services/ai
@@ -1,1 +1,0 @@
-../../workers/ai
