+++ doctor-who-ai/app/workers/ai/rnn/tf-utils.ts
@@ -1,0 +1,56 @@
+import * as tf from '@tensorflow/tfjs';
+
+const fileName = 'guided-medium.model';
+const dataLocation = `downloads://${fileName}`;
+const fileInfoLocation = `http://localhost:4200/${fileName}.json`;
+
+export async function useGPU() {
+  if (tf.getBackend() !== 'webgl') {
+    await tf.setBackend('webgl');
+  }
+
+  await tf.ready();
+}
+
+export async function save(network: tf.LayersModel) {
+  await network.save(dataLocation);
+}
+
+export async function getNetwork(): Promise<tf.LayersModel> {
+  let model;
+
+  try {
+    model = await tf.loadLayersModel(fileInfoLocation);
+  } catch (e) {
+    console.debug(e);
+  }
+
+  if (!model) {
+    model = createNetwork();
+  }
+
+  model.compile({ optimizer: 'sgd', loss: 'meanSquaredError' });
+  model.summary();
+
+  return model;
+}
+
+function createNetwork() {
+  let layer = tf.layers.dense;
+
+  let model = tf.sequential({
+    name: '2048-move-network',
+    layers: [
+      layer({ name: 'input', units: Math.pow(2, 9), inputShape: [16], activation: 'relu' }),
+      layer({ name: 'hidden-1', units: Math.pow(2, 11), activation: 'relu' }),
+      layer({ name: 'hidden-2', units: Math.pow(2, 9), activation: 'relu' }),
+      layer({ name: 'hidden-3', units: Math.pow(2, 8), activation: 'relu' }),
+      layer({ name: 'hidden-4', units: Math.pow(2, 7), activation: 'relu' }),
+      layer({ name: 'hidden-5', units: Math.pow(2, 6), activation: 'relu' }),
+      layer({ name: 'hidden-6', units: Math.pow(2, 5), activation: 'relu' }),
+      layer({ name: 'output', units: 4, activation: 'softmax' }),
+    ],
+  });
+
+  return model;
+}
