+++ doctor-who-ai/package.json
@@ -23,8 +23,19 @@
     "test:ember": "ember test"
   },
   "dependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.10.4",
+    "@babel/plugin-proposal-object-rest-spread": "^7.10.4",
+    "@babel/plugin-transform-typescript": "^7.10.5",
+    "@babel/preset-env": "^7.10.4",
+    "@babel/preset-typescript": "^7.10.4",
+    "@rollup/plugin-babel": "5.1.0",
+    "@rollup/plugin-commonjs": "14.0.0",
+    "@rollup/plugin-node-resolve": "7.1.3",
+    "@skyrocketjs/ember": "0.0.1-alpha.1",
+    "@skyrocketjs/service": "0.0.1-alpha.1",
     "@tensorflow/tfjs": "2.0.1",
     "broccoli-merge-trees": "*",
+    "broccoli-rollup": "4.1.1",
     "broccoli-source": "*",
     "chartist-plugin-legend": "0.6.2",
     "ember-cli-chartist": "3.0.0",
@@ -33,10 +44,12 @@
     "ember-concurrency-decorators": "2.0.1",
     "ember-concurrency-ts": "0.2.0",
     "git-rev-sync": "2.0.0",
-    "random": "2.2.0"
-  },
-  "resolutions": {
-    "@tensorflow/tfjs": "2.0.1"
+    "promise-worker-bi": "4.0.2",
+    "random": "2.2.0",
+    "rollup": "2.23.0",
+    "core-js": "3.6.5",
+    "rollup-plugin-filesize": "9.0.2",
+    "rollup-plugin-terser": "6.1.0"
   },
   "devDependencies": {
     "@ember/optional-features": "^1.3.0",
@@ -75,7 +88,6 @@
     "eslint-plugin-ember": "^8.9.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "lint-staged": "^10.2.11",
     "loader.js": "^4.7.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.0.5",
