--- doctor-who-ai/workers/ai/tfjs-mountaincar/README.md
@@ -1,5 +1,0 @@
-Code originally from
-
-https://github.com/prouhard/tfjs-mountaincar
-
-https://medium.com/@pierrerouhard/reinforcement-learning-in-the-browser-an-introduction-to-tensorflow-js-9a02b143c099
